fix: preserve push messages in cache until they are seen by client (#108)
(#103)

Atmosphere caches messages when the client is disconnected, but
unfortunately it may happen that a message does not reach the client
because of network disconnection during async response write operation.
In this case the message is not cached and will be completely lost,
causing a UI resynchronization request.
This change preserves messages in broadcaster cache until the client
confirms that they have been processed, by sending the last seen server
sync identifier on reconnection.
This should prevent the need for a UI resynchronization. It may happen
in some cases, e.g. back to online after being offline, that messages
already seen will be sent to the client, but those will be discarded.

Backported from Flow: https://github.com/vaadin/flow/pull/15764